+++ cryptofeed/standards.py
@@ -12,13 +12,12 @@ data channel names
 import logging
 import datetime as dt
 
-from cryptofeed.defines import (BINANCE, BINANCE_DELIVERY, BINANCE_FUTURES, BINANCE_US, BITCOINCOM, BITFLYER, BITFINEX,
-                                BITHUMB, BITMAX, BITMEX,
-                                BITSTAMP, BITTREX, BLOCKCHAIN, BYBIT, CANDLES, COINBASE, COINGECKO,
+from cryptofeed.defines import (BEQUANT, BINANCE, BINANCE_DELIVERY, BINANCE_FUTURES, BINANCE_US, BITCOINCOM, BITFLYER, BITFINEX,
+                                BITHUMB, ASCENDEX, BITMEX, PHEMEX, BITSTAMP, BITTREX, BLOCKCHAIN, BYBIT, CANDLES, COINBASE, COINGECKO,
                                 DERIBIT, DYDX, EXX, FTX, FTX_US, GATEIO, GEMINI, HITBTC, HUOBI, HUOBI_DM, HUOBI_SWAP,
-                                KRAKEN, KRAKEN_FUTURES, KUCOIN, OKCOIN, OKEX, PHEMEX, POLONIEX, PROBIT, UPBIT, USER_FILLS)
+                                KRAKEN, KRAKEN_FUTURES, KUCOIN, OKCOIN, OKEX, POLONIEX, PROBIT, ACC_TRANSACTIONS, UPBIT, USER_FILLS)
 from cryptofeed.defines import (FILL_OR_KILL, IMMEDIATE_OR_CANCEL, LIMIT, MAKER_OR_CANCEL, MARKET, UNSUPPORTED)
-from cryptofeed.defines import (FUNDING, FUTURES_INDEX, L2_BOOK, L3_BOOK, LIQUIDATIONS, OPEN_INTEREST, MARKET_INFO,
+from cryptofeed.defines import (ACC_BALANCES, FUNDING, FUTURES_INDEX, L2_BOOK, L3_BOOK, LIQUIDATIONS, OPEN_INTEREST, MARKET_INFO,
                                 TICKER, TRADES, ORDER_INFO)
 from cryptofeed.exceptions import UnsupportedDataFeed, UnsupportedTradingOption
 
@@ -32,10 +31,10 @@ def timestamp_normalize(exchange, ts):
             return ts / 1000
         else:
             return ts.timestamp()
-    if exchange in {BITFLYER, COINBASE, BLOCKCHAIN, BITMEX, HITBTC, OKCOIN, OKEX, FTX, FTX_US, BITCOINCOM, PROBIT, COINGECKO, BITTREX, DYDX}:
+    if exchange in {BITFLYER, COINBASE, BLOCKCHAIN, BITMEX, HITBTC, OKCOIN, OKEX, FTX, FTX_US, BITCOINCOM, PROBIT, COINGECKO, BITTREX, DYDX, BEQUANT}:
         return ts.timestamp()
     elif exchange in {HUOBI, HUOBI_DM, HUOBI_SWAP, BITFINEX, DERIBIT, BINANCE, BINANCE_US, BINANCE_FUTURES,
-                      BINANCE_DELIVERY, GEMINI, BITMAX, KRAKEN_FUTURES, UPBIT}:
+                      BINANCE_DELIVERY, GEMINI, ASCENDEX, KRAKEN_FUTURES, UPBIT}:
         return ts / 1000.0
     elif exchange in {BITSTAMP}:
         return ts / 1000000.0
@@ -47,6 +46,7 @@ def timestamp_normalize(exchange, ts):
 _feed_to_exchange_map = {
     L2_BOOK: {
         DYDX: 'v3_orderbook',
+        BEQUANT: 'subscribeOrderbook',
         BITFINEX: 'book-P0-F0-100',
         BITFLYER: 'lightning_board_{}',
         BITHUMB: 'orderbookdepth',
@@ -75,7 +75,7 @@ _feed_to_exchange_map = {
         GEMINI: L2_BOOK,
         BITTREX: 'orderbook_{}_{}',
         BITCOINCOM: 'subscribeOrderbook',
-        BITMAX: "depth:",
+        ASCENDEX: "depth:",
         UPBIT: L2_BOOK,
         GATEIO: 'spot.order_book_update',
         PROBIT: 'order_books',
@@ -83,6 +83,7 @@ _feed_to_exchange_map = {
         PHEMEX: 'orderbook.subscribe'
     },
     L3_BOOK: {
+        BEQUANT: UNSUPPORTED,
         BITTREX: UNSUPPORTED,
         BITFINEX: 'book-R0-F0-100',
         BITHUMB: UNSUPPORTED,
@@ -108,12 +109,13 @@ _feed_to_exchange_map = {
         FTX_US: UNSUPPORTED,
         GEMINI: UNSUPPORTED,
         BITCOINCOM: UNSUPPORTED,
-        BITMAX: UNSUPPORTED,
+        ASCENDEX: UNSUPPORTED,
         UPBIT: UNSUPPORTED,
         PROBIT: UNSUPPORTED
     },
     TRADES: {
         DYDX: 'v3_trades',
+        BEQUANT: 'subscribeTrades',
         POLONIEX: TRADES,
         HITBTC: 'subscribeTrades',
         BITSTAMP: 'live_trades',
@@ -142,7 +144,7 @@ _feed_to_exchange_map = {
         GEMINI: TRADES,
         BITTREX: 'trade_{}',
         BITCOINCOM: 'subscribeTrades',
-        BITMAX: "trades:",
+        ASCENDEX: "trades:",
         UPBIT: TRADES,
         GATEIO: 'spot.trades',
         PROBIT: 'recent_trades',
@@ -150,6 +152,7 @@ _feed_to_exchange_map = {
         PHEMEX: 'trade.subscribe'
     },
     TICKER: {
+        BEQUANT: 'subscribeTicker',
         POLONIEX: 1002,
         HITBTC: 'subscribeTicker',
         BITFINEX: 'ticker',
@@ -175,7 +178,7 @@ _feed_to_exchange_map = {
         GEMINI: UNSUPPORTED,
         BITTREX: 'ticker_{}',
         BITCOINCOM: 'subscribeTicker',
-        BITMAX: UNSUPPORTED,
+        ASCENDEX: UNSUPPORTED,
         UPBIT: UNSUPPORTED,
         GATEIO: 'spot.tickers',
         PROBIT: UNSUPPORTED,
@@ -221,11 +224,13 @@ _feed_to_exchange_map = {
         GEMINI: ORDER_INFO,
         OKEX: ORDER_INFO,
         FTX: 'orders',
+        BEQUANT: 'subscribeReports',
     },
     USER_FILLS: {
         FTX: 'fills',
     },
     CANDLES: {
+        BEQUANT: 'subscribeCandles',
         BINANCE: 'kline_',
         BINANCE_US: 'kline_',
         BINANCE_FUTURES: 'kline_',
@@ -236,11 +241,18 @@ _feed_to_exchange_map = {
         KRAKEN: 'ohlc',
         BITTREX: 'candle_{}_{}',
         PHEMEX: 'kline.subscribe'
-    }
+    },
+    ACC_TRANSACTIONS: {
+        BEQUANT: 'subscribeTransactions',
+    },
+    ACC_BALANCES: {
+        BEQUANT: 'subscribeBalance'
+    },
 }
 
 _exchange_options = {
     LIMIT: {
+        BEQUANT: 'limit',
         KRAKEN: 'limit',
         GEMINI: 'exchange limit',
         POLONIEX: 'limit',
@@ -248,6 +260,7 @@ _exchange_options = {
         BLOCKCHAIN: 'limit',
     },
     MARKET: {
+        BEQUANT: 'market',
         KRAKEN: 'market',
         GEMINI: UNSUPPORTED,
         POLONIEX: UNSUPPORTED,
@@ -255,6 +268,7 @@ _exchange_options = {
         BLOCKCHAIN: 'market',
     },
     FILL_OR_KILL: {
+        BEQUANT: {'timeInForce': 'FOK'},
         GEMINI: 'fill-or-kill',
         POLONIEX: 'fillOrKill',
         COINBASE: {'time_in_force': 'FOK'},
@@ -262,6 +276,7 @@ _exchange_options = {
         BLOCKCHAIN: 'FOK'
     },
     IMMEDIATE_OR_CANCEL: {
+        BEQUANT: {'timeInForce': 'IOC'},
         GEMINI: 'immediate-or-cancel',
         POLONIEX: 'immediateOrCancel',
         COINBASE: {'time_in_force': 'IOC'},
@@ -269,6 +284,7 @@ _exchange_options = {
         BLOCKCHAIN: 'IOC'
     },
     MAKER_OR_CANCEL: {
+        BEQUANT: {'postOnly': 1},
         GEMINI: 'maker-or-cancel',
         POLONIEX: 'postOnly',
         COINBASE: {'post_only': 1},
@@ -315,4 +331,4 @@ def normalize_channel(exchange: str, feed: str) -> str:
 
 
 def is_authenticated_channel(channel: str) -> bool:
-    return channel in (ORDER_INFO, USER_FILLS)
+    return channel in (ORDER_INFO, USER_FILLS, ACC_TRANSACTIONS, ACC_BALANCES)