+++ CHANGES.md
@@ -18,6 +18,9 @@
   * Feature: Allow user to override the score used in Redis ZSETs
   * Update: Get information about size increment from FTX symbol data
   * Bugfix: Fix trades write for Arctic backend
+  * Feature: new exchange: Bequant. Supports ticker, L2 book, trades, candles, plus authenticated channels: order info, account transactions and account balances
+  * Update: BitMax renamed AscendEX
+  * Bugfix: Feed level timeout and timeout interval not being set properly
 
 ### 1.9.1 (2021-06-10)
   * Feature: add Bithumb exchange - l2 book and trades